+++ CHANGELOG.md
@@ -6,6 +6,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [6.0.0]
+### Updated
+- Support for 1.19.2
+- Addons: [Applied Energistics], [Ars Nouveau], [Blue Skies], [Extended Crafting], [Mystical Agriculture]
+
+### Removed
+- Support for [Botania], [Chipped], [Create], [Immersive Engineering], [Mana and Artifice], Pam's Harvest Craft, Thermal, [Touhou Little Maid]
+- DataGen for [Blue Skies]
+
+## [5.3.2]
+### Added
+- Partial Support for [Blue Skies] recipes and emc values
+
+### Changed
+- Using Parchment mappings
+
 ## [5.3.1]
 ### Added
 - Support for [Ars Nouveau] recipes and emc to items
@@ -601,7 +617,9 @@ Added support for 18 more mods
 - Added support for Extended Crafting
     - Combintation Core, Compressor, Ender Crafter and Tiered Tables
 
-[Unreleased]: https://github.com/TagnumElite/ProjectE-Integration/compare/v5.3.1...HEAD
+[Unreleased]: https://github.com/TagnumElite/ProjectE-Integration/compare/v6.0.0...HEAD
+[6.0.0]: https://github.com/TagnumElite/ProjectE-Integration/compare/v5.3.2...v6.0.0
+[5.3.2]: https://github.com/TagnumElite/ProjectE-Integration/compare/v5.3.1...v5.3.2
 [5.3.1]: https://github.com/TagnumElite/ProjectE-Integration/compare/v5.3.0...v5.3.1
 [5.3.0]: https://github.com/TagnumElite/ProjectE-Integration/compare/v5.2.4...v5.3.0
 [5.2.4]: https://github.com/TagnumElite/ProjectE-Integration/compare/v5.2.3...v5.2.4